+++ web/server.js
@@ -78,21 +78,31 @@ router.get('/api/strategies', require(ROUTE('strategies')));
 router.get('/api/configPart/:part', require(ROUTE('configPart')));
 router.get('/api/apiKeys', apiKeys.get);
 
+router.get('/api/backtests', require(ROUTE('backtests')));
+router.get('/api/backtests/:id', require(ROUTE('backtestResult')));
+router.delete('/api/backtests/:id', require(ROUTE('backtestDeleteResult')));
+
 const listWraper = require(ROUTE('list'));
 router.get('/api/imports', listWraper('imports'));
 router.get('/api/gekkos', listWraper('gekkos'));
 router.get('/api/exchanges', require(ROUTE('exchanges')));
+
 
 router.post('/api/addApiKey', apiKeys.add);
 router.post('/api/removeApiKey', apiKeys.remove);
 router.post('/api/scan', require(ROUTE('scanDateRange')));
 router.post('/api/scansets', require(ROUTE('scanDatasets')));
 router.post('/api/backtest', require(ROUTE('backtest')));
-router.post('/api/imports', require(ROUTE('import')));
 router.post('/api/startGekko', require(ROUTE('startGekko')));
 router.post('/api/stopGekko', require(ROUTE('stopGekko')));
 router.post('/api/deleteGekko', require(ROUTE('deleteGekko')));
 router.post('/api/getCandles', require(ROUTE('getCandles')));
+
+const imports = require(ROUTE('import'));
+
+// Keep legacy ui working
+router.post('/api/import', imports);
+router.post('/api/imports', imports);
 
 // Sandbox API
 router.post('/api/sandboxBacktest', require(ROUTE('sandboxBacktest')));